orbitalskyshield/pyproject.toml
[project]
name = "orbitalskyshield"
version = "0.1.0"
description = "Satellite streak detection and orbital diffuse contribution estimation."
authors = [
    {name = "Antigravity", email = "antigravity@example.com"},
]
dependencies = [
    "astropy>=5.0",
    "numpy>=1.20",
    "scipy>=1.7",
    "matplotlib>=3.5",
    "pyyaml>=6.0",
    "typer>=0.9",
    "scikit-image>=0.19",
    "pandas>=1.4",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.0.260",
    "mypy>=1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["orbitalskyshield*"]
